Reshuffle some CDC things, make Xilinx CDC defs into generic ones
NOTE: will not pass CI
